+++ src/errors/rejectOnCancelled.ts
@@ -1,0 +1,44 @@
+/*
+ * Copyright (C) 2022 PixieBrix, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import { PromiseCancelled } from "@/errors/genericErrors";
+
+/**
+ * Creates a new promise that's rejected if isCancelled returns true.
+ * @throws PromiseCancelled
+ */
+export async function rejectOnCancelled<T>(
+  promise: Promise<T>,
+  isCancelled: () => boolean
+): Promise<T> {
+  let rv: T;
+  try {
+    rv = await promise;
+  } catch (error) {
+    if (isCancelled()) {
+      throw new PromiseCancelled();
+    }
+
+    throw error ?? new Error("Undefined error awaiting promise");
+  }
+
+  if (isCancelled()) {
+    throw new PromiseCancelled();
+  }
+
+  return rv;
+}
